+++ src/main/java/com/tiviacz/travelersbackpack/network/ServerboundSlotPacket.java
@@ -17,16 +17,16 @@ import java.util.function.Supplier;
 import java.util.stream.Collectors;
 
 public class ServerboundSlotPacket {
-    private final byte selectType;
+    private final int selectType;
     private final List<?> slotsData;
 
-    public ServerboundSlotPacket(byte selectType, List<?> slotsData) {
+    public ServerboundSlotPacket(int selectType, List<?> slotsData) {
         this.selectType = selectType;
         this.slotsData = slotsData;
     }
 
     public static ServerboundSlotPacket decode(final FriendlyByteBuf buffer) {
-        final byte selectType = buffer.readByte();
+        final int selectType = buffer.readInt();
         List<?> slotsData = new ArrayList<>();
         if(selectType == UNSORTABLES) {
             slotsData = buffer.readIntIdList().intStream().boxed().collect(Collectors.toList());
@@ -38,7 +38,7 @@ public class ServerboundSlotPacket {
     }
 
     public static void encode(final ServerboundSlotPacket message, final FriendlyByteBuf buffer) {
-        buffer.writeByte(message.selectType);
+        buffer.writeInt(message.selectType);
         List<?> slotsData = message.slotsData;
         if(message.selectType == UNSORTABLES) {
             List<Integer> unsortables = (List<Integer>)slotsData;
@@ -49,8 +49,8 @@ public class ServerboundSlotPacket {
         }
     }
 
-    public static final byte UNSORTABLES = (byte)0;
-    public static final byte MEMORY = (byte)1;
+    public static final int UNSORTABLES = 0;
+    public static final int MEMORY = 1;
 
     public static void handle(final ServerboundSlotPacket message, final Supplier<NetworkEvent.Context> ctx) {
         ctx.get().enqueueWork(() -> {
